Spread the query builder over two modules
One for the metaprogramming, one for interpreting the query at run time.
A new context object has been introduced, which makes it easier to pass
common data around.

The QueryBuilder module is now just a place for callbacks and
delegate-like functions. The goal behind this is to keep the
implementation elsewhere, while the docs can live in the query builder.